Fixed and fully implemented splash screen.
