+++ lib/splash_screen.dart
@@ -1,0 +1,64 @@
+import 'package:flutter/material.dart';
+import 'package:restaurant_reservation_app/prototype_tester.dart';
+import 'custom_color_swatch.dart';
+
+void main() => runApp(const SplashScreen());
+
+class SplashScreen extends StatelessWidget {
+  const SplashScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Prototype Tester',
+      theme: ThemeData(
+        primarySwatch: CustomColorSwatch.swatch,
+      ),
+      home: const Scaffold(
+        body: SplashScreenPage(),
+      ),
+    );
+  }
+}
+
+class SplashScreenPage extends StatefulWidget {
+  const SplashScreenPage({Key? key}) : super(key: key);
+
+  @override
+  State<SplashScreenPage> createState() => _SplashScreenPageState();
+}
+
+class _SplashScreenPageState extends State<SplashScreenPage> {
+  @override
+  void initState() {
+    super.initState();
+    navigateToHome();
+  }
+
+  navigateToHome() async {
+    await Future.delayed(const Duration(milliseconds: 3000), () {});
+    Navigator.pushReplacement(
+        context,
+        MaterialPageRoute(builder: (context) => const PrototypeTester()));
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Container(
+              height: 100,
+              width: 100,
+              color: Colors.blue,
+            ),
+            const Text('Restaurant Reservation App',
+                style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold)),
+          ],
+        ),
+      ),
+    );
+  }
+}
